BaseCharacter: add SetWeaponCollisionEnabled
