+++ Source/Delta/Character/BaseCharacter.cpp
@@ -7,9 +7,24 @@
 #include "Components/CapsuleComponent.h"
 #include "GameFramework/SpringArmComponent.h"
 #include "Camera/CameraComponent.h"
+#include "Components/BoxComponent.h"
+#include "../Common/Finder.h"
+#include "../Common/LogUtil.h"
+#include "../Weapon/Weapon.h"
+#include "../Weapon/Sword.h"
 
 ABaseCharacter::ABaseCharacter() {
-  PrimaryActorTick.bCanEverTick = true;
+  {
+    PrimaryActorTick.bCanEverTick = true;
+
+    bUseControllerRotationPitch = false;
+    bUseControllerRotationYaw   = false;
+    bUseControllerRotationRoll  = false;
+  }
+ 
+  {
+
+  }
 }
 
 void ABaseCharacter::BeginPlay() {
@@ -22,4 +37,17 @@ void ABaseCharacter::Tick(float DeltaTime) {
 
 void ABaseCharacter::SetupPlayerInputComponent(UInputComponent* PlayerInputComponent) {
   Super::SetupPlayerInputComponent(PlayerInputComponent);
+}
+
+void ABaseCharacter::SetWeaponCollisionEnabled(ECollisionEnabled::Type CollisionEnabled) {
+  if (EquippedWeapon == nullptr) {
+    return;
+  }
+  EquippedWeapon->IgnoreActors.Empty();
+
+  auto* const WeaponBox = EquippedWeapon->GetWeaponBox();
+  if (WeaponBox == nullptr) {
+    return;
+  }
+  WeaponBox->SetCollisionEnabled(CollisionEnabled);
 }
